se arreglo cohortes, se modifico ruta eliminar cohorte por id, herramientas/alumno pega con db
